requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
redis>=5.0.0
pybaseball>=2.3.0
beautifulsoup4>=4.12.0
requests>=2.31.0
httpx>=0.25.0
pandas>=2.1.0
numpy>=1.26.0
openpyxl>=3.1.0
apscheduler>=3.10.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
